Added and packages files: -Control.java -DatabaseSQL.DatabaseConnection.java -GUI.ConnectionGUI.java and ConnectionGUI.form
Added 4 icons for the GUI

What has been added:
A GUI has been created which allows the user to enter details which is used to connect to a postgres database. JDBC is used to connect to the postgres database. The GUI will tell the user whether or not the connections was successful.