Added Files: - GuiControl.java - FileGui.java - FileGui.form
Changed File:
- ConnectionGui.java
- ConnectionGui.form.
These files used to have a different function but now that has been changed. The GuiControl.java file controls all the aspects of the GUI for this program. ConnectionGui.java and FileGui.java only purpose is to create the GUI for this program and to set and return values using getters and setters.

Comments for every file has been changed and improved.
Some new private variables have been created.
Added a new CSV file icon for the GUI.